fixed bold font case not showing
